Change LIN to LUM in sector classification table
It seems to me this is a typo with wood products carrying the same CGE code as textiles. However, if my assumption is wrong and LIN is correct, then I would suggest to adjust the formatting of the table such that in the CGE code column, the two rows are joined to signal that this is intentional.